+++ http/listAccounts.go
@@ -14,12 +14,12 @@ type ListAccountResponse struct {
 	Id        string    `json:"id"`
 	Name      string    `json:"name"`
 	Cpf       int       `json:"cpf"`
-	Balance   float64   `json:"balance"`
+	Balance   int64     `json:"balance"`
 	CreatedAt time.Time `json:"createdAt"`
 }
 
 type BalanceAccountResponse struct {
-	Balance float64 `json:"balance"`
+	Balance int64 `json:"balance"`
 }
 
 type AccountIdRequest struct {
